make scopes optional
If both `scope` and `user_scope` are provided, Slack will return a
bot token instead of a user token. Previously, setting the
`default_scope` to the empty string was causing issues as the
authorize URL was trying to encode it weirdly and blowing up. Now
we only include the `scope` parameter if it is *not* the empty string.